+++ Source/ACE.Server/Network/Enum/GenericQualitiesPackHeader.cs
@@ -1,0 +1,11 @@
+namespace ACE.Server.Network.Enum
+{
+    public enum GenericQualitiesPackHeader
+    {
+        Packed_None = 0,
+        Packed_IntStats = 1,
+        Packed_BoolStats = 2,
+        Packed_FloatStats = 4,
+        Packed_StringStats = 8,
+    };
+}